# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace API",
    "version": "1.0.0",
    "description": "REST API voor een marketplace met users, skins, listings en transactions."
  },
  "servers": [
    {
      "url": "https://marketplace-xbmm.onrender.com",
      "description": "Live Render server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Lokale development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-token"
      }
    },
    "schemas": {
...
      "ListingRequest": {
        "type": "object",
        "required": ["price", "skinId"],
        "properties": {
          "price": { "type": "number", "example": 350 },
          "skinId": { "type": "string", "example": "665f2d4b4c6a8a1b2c3d4e5f" }
        }
      },
      "ListingUpdateRequest": {
        "type": "object",
        "required": ["price"],
        "properties": {
          "price": { "type": "number", "example": 400 },
          "status": {
            "type": "string",
            "enum": ["active", "sold", "cancelled"],
            "example": "active"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/listings": {
      "get": {
        "summary": "Alle actieve listings ophalen",
        "tags": ["Listings"],
        "responses": {
          "200": { "description": "Lijst van actieve listings" }
        }
      },
      "post": {
        "summary": "Nieuwe listing maken",
        "tags": ["Listings"],
        "security": [{ "AuthToken": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListingRequest" } } }
        },
        "responses": {
          "201": { "description": "Listing aangemaakt" },
          "401": { "description": "Niet ingelogd" }
        }
      }
    },
    "/api/listings/search": {
      "get": {
        "summary": "Listings zoeken",
        "tags": ["Listings"],
        "parameters": [
          { "name": "minPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" } },
          {
            "name": "rarity",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["Common", "Uncommon", "Rare", "Epic", "Legendary"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Gefilterde actieve listings" },
          "400": { "description": "Ongeldige query parameter" }
        }
      }
    },
    "/api/listings/my-listings": {
      "get": {
        "summary": "Eigen listings ophalen",
        "tags": ["Listings"],
        "security": [{ "AuthToken": [] }],
        "responses": {
          "200": { "description": "Listings van ingelogde gebruiker" },
          "401": { "description": "Niet ingelogd" }
        }
      }
    },
    "/api/listings/{id}": {
      "get": {
        "summary": "Listing ophalen op id",
        "tags": ["Listings"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Listing gevonden" },
          "400": { "description": "Ongeldig ObjectId" },
          "404": { "description": "Listing niet gevonden" }
        }
      },
      "put": {
        "summary": "Listing aanpassen",
        "tags": ["Listings"],
        "security": [{ "AuthToken": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListingUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "Listing aangepast" },
          "403": { "description": "Geen toegang" },
          "404": { "description": "Listing niet gevonden" }
        }
      },
      "delete": {
        "summary": "Listing verwijderen",
        "tags": ["Listings"],
        "security": [{ "AuthToken": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Listing verwijderd" },
          "403": { "description": "Geen toegang" },
          "404": { "description": "Listing niet gevonden" }
        }
      }
    },
    "/api/listings/{id}/buy": {
      "post": {
        "summary": "Listing kopen",
        "tags": ["Listings"],
        "security": [{ "AuthToken": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Aankoop succesvol" },
          "400": { "description": "Onvoldoende credits of listing niet actief" },
          "404": { "description": "Listing niet gevonden" }
        }
      }
    },
...
  }
}